sankalps_intelli_layer/AI-BEING-main/Backend Stability Spine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.2
openai==1.3.0
groq==0.4.1
google-generativeai==0.3.2
mistralai==0.4.2
requests==2.31.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.27.0
numpy==1.26.4
sqlalchemy==2.0.23
aiosqlite==0.19.0
openai-whisper==20231117
torch==2.2.0
transformers==4.35.2
sentence-transformers==2.2.2
notion-client
gspread
trello
oauth2client
sentry-sdk[fastapi]==1.40.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
python-dateutil==2.8.2
